Делаю приведения типа для модуля в макросе P7_LOG, чтобы его легче было использовать.

+++ Singleton/p7_logger.h
@@ -78,7 +78,7 @@ private:
 #define P7_LOG(log_level, module_enum, format, ...)\
    p7_logger_raii::instance().trace().P7_DELIVER(0,\
                                             static_cast<size_t>(log_level),\
-                                            p7_logger_raii::instance().module(module_enum),\
+                                            p7_logger_raii::instance().module(static_cast<size_t>(module_enum)),\
                                             format,\
                                             ##__VA_ARGS__)
 #else // USE_P7_LOG
